Good intermediate point, refactored schema in its own crate
