+++ core/cu29/Cargo.toml
@@ -22,6 +22,7 @@ cu29-unifiedlog = { workspace = true }
 cu29-clock = { workspace = true }
 cu29-value = { workspace = true }
 cu29-intern-strs = { workspace = true }
+cu29-schema = { workspace = true }
 bincode = { workspace = true }
 
 [features]
