+++ core/cu29_clock/Cargo.toml
@@ -15,6 +15,7 @@ repository.workspace = true
 quanta = "0.12.5"
 bincode = { workspace = true }
 serde = { workspace = true }
+cu29-schema = { workspace = true }
 
 [dev-dependencies]
 approx = "0.5.1"
