+++ core/cu29_runtime/Cargo.toml
@@ -28,6 +28,7 @@ cu29-log-runtime = { workspace = true }  # needed
 cu29-unifiedlog = { workspace = true }
 cu29-value = { workspace = true }
 cu29-clock = { workspace = true }
+cu29-schema = { workspace = true }
 clap = { workspace = true }
 tempfile = { workspace = true }
 arrayvec = "0.7.6"
