+++ core/cu29_schema/Cargo.toml
@@ -1,0 +1,12 @@
+[package]
+name = "cu29-schema"
+version.workspace = true
+authors.workspace = true
+edition.workspace = true
+license.workspace = true
+keywords.workspace = true
+categories.workspace = true
+homepage.workspace = true
+repository.workspace = true
+
+[dependencies]
