+++ core/cu29_schema/src/lib.rs
@@ -1,0 +1,328 @@
+//! Schema trait and types for copper runtime
+//! 
+//! This crate provides the Schema trait and SchemaType enum for describing
+//! the structure of data types in the copper runtime system.
+
+use std::collections::HashMap;
+use std::fmt::{Display, Formatter};
+
+/// Represents the type information for schema generation
+#[derive(Debug, Clone, PartialEq)]
+pub enum SchemaType {
+    U8,
+    U16,
+    U32,
+    U64,
+    U128,
+    I8,
+    I16,
+    I32,
+    I64,
+    I128,
+    F32,
+    F64,
+    Bool,
+    String,
+    Vec(Box<SchemaType>),
+    Option(Box<SchemaType>),
+    Tuple(Vec<SchemaType>),
+    Custom(String), // For foreign structs that don't implement Schema
+    Struct {
+        name: String,
+        fields: HashMap<String, SchemaType>,
+    }, // For structs that implement Schema and allow recursive inspection
+}
+
+impl Display for SchemaType {
+    fn fmt(&self, f: &mut Formatter<'_>) -> std::fmt::Result {
+        match self {
+            SchemaType::U8 => write!(f, "u8"),
+            SchemaType::U16 => write!(f, "u16"),
+            SchemaType::U32 => write!(f, "u32"),
+            SchemaType::U64 => write!(f, "u64"),
+            SchemaType::U128 => write!(f, "u128"),
+            SchemaType::I8 => write!(f, "i8"),
+            SchemaType::I16 => write!(f, "i16"),
+            SchemaType::I32 => write!(f, "i32"),
+            SchemaType::I64 => write!(f, "i64"),
+            SchemaType::I128 => write!(f, "i128"),
+            SchemaType::F32 => write!(f, "f32"),
+            SchemaType::F64 => write!(f, "f64"),
+            SchemaType::Bool => write!(f, "bool"),
+            SchemaType::String => write!(f, "String"),
+            SchemaType::Vec(inner) => write!(f, "Vec<{inner}>"),
+            SchemaType::Option(inner) => write!(f, "Option<{inner}>"),
+            SchemaType::Tuple(types) => {
+                write!(f, "(")?;
+                for (i, ty) in types.iter().enumerate() {
+                    if i > 0 {
+                        write!(f, ", ")?;
+                    }
+                    write!(f, "{}", ty)?;
+                }
+                write!(f, ")")
+            }
+            SchemaType::Custom(name) => write!(f, "Custom({name})"),
+            SchemaType::Struct { name, .. } => write!(f, "Struct {name}"),
+        }
+    }
+}
+
+/// Trait for types that can provide schema information
+pub trait Schema {
+    /// Returns a HashMap mapping field names to their types
+    fn schema() -> HashMap<String, SchemaType>;
+
+    /// Returns the type name of the struct
+    fn type_name() -> &'static str;
+
+    /// Returns the schema as a SchemaType::Struct for recursive inspection
+    fn schema_type() -> SchemaType {
+        SchemaType::Struct {
+            name: Self::type_name().to_string(),
+            fields: Self::schema(),
+        }
+    }
+
+    /// Prints the schema in a human-readable format
+    fn print_schema() {
+        let schema = Self::schema();
+        let type_name = Self::type_name();
+        println!("Schema for {type_name}:");
+        for (field_name, field_type) in schema {
+            println!("  {field_name}: {field_type}");
+        }
+    }
+}
+
+impl Schema for () {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "(empty)"
+    }
+}
+
+// Schema implementations for primitive types
+impl Schema for u8 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "u8"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::U8
+    }
+}
+
+impl Schema for u16 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "u16"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::U16
+    }
+}
+
+impl Schema for u32 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "u32"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::U32
+    }
+}
+
+impl Schema for u64 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "u64"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::U64
+    }
+}
+
+impl Schema for u128 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "u128"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::U128
+    }
+}
+
+impl Schema for i8 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "i8"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::I8
+    }
+}
+
+impl Schema for i16 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "i16"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::I16
+    }
+}
+
+impl Schema for i32 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "i32"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::I32
+    }
+}
+
+impl Schema for i64 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "i64"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::I64
+    }
+}
+
+impl Schema for i128 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "i128"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::I128
+    }
+}
+
+impl Schema for f32 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "f32"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::F32
+    }
+}
+
+impl Schema for f64 {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "f64"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::F64
+    }
+}
+
+impl Schema for bool {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "bool"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::Bool
+    }
+}
+
+impl Schema for String {
+    fn schema() -> HashMap<String, SchemaType> {
+        HashMap::new()
+    }
+
+    fn type_name() -> &'static str {
+        "String"
+    }
+
+    fn schema_type() -> SchemaType {
+        SchemaType::String
+    }
+}
+
+macro_rules! impl_schema_for_tuples {
+    ($(($($ty:ident),*)),*) => {
+        $(
+            impl<$($ty: Schema),*> Schema for ($($ty,)*) {
+                fn schema() -> HashMap<String, SchemaType> {
+                    HashMap::new()
+                }
+
+                fn type_name() -> &'static str {
+                    "tuple"
+                }
+
+                fn schema_type() -> SchemaType {
+                    SchemaType::Tuple(vec![$($ty::schema_type()),*])
+                }
+            }
+        )*
+    };
+}
+
+// Apply the macro to generate Schema implementations for tuple sizes up to 5
+impl_schema_for_tuples! {
+    (T1, T2), (T1, T2, T3), (T1, T2, T3, T4), (T1, T2, T3, T4, T5)
+}
